add serde check to pytest framework
