+++ tests/python/test_python_frontend.py
@@ -8,7 +8,7 @@ from functools import partial
 import itertools
 import math
 import random
-from typing import List
+from typing import List, Callable
 import unittest
 
 import torch
@@ -34,11 +34,52 @@ from utils import (
     is_pre_ampere,
     is_pre_hopper,
     check_captured_python_definition,
-    serde_check,
+    basic_serde_check,
     debug_serde,
 )
 
 RUN_NVFUSER = RUN_CUDA and not TEST_WITH_ROCM
+
+
+def serde_check(test_fn: Callable):
+    """
+    A decorator to verify that serialization works with the given exec_nvfuser
+    function. Currently, it uses serialization to rebuild the FusionCache
+    structure.
+    """
+
+    def inner_fn(*args, **kwargs):
+        self, fusion_func, inputs = args
+
+        # NOTE: For debug purposes, clear FusionCache before running first test
+        # so the behavior is more deterministic (PR #1848).
+        is_new_fusion_expected = kwargs.get("new_fusion_expected", True)
+        if debug_serde and is_new_fusion_expected:
+            FusionCache.reset()
+            assert FusionCache.get().num_fusions() == 0
+
+        # skip_serde_check is only used by the decorator so remove it before
+        # running test_fn
+        skip_serde_check = kwargs.pop("skip_serde_check", False)
+        if skip_serde_check:
+            return test_fn(self, fusion_func, inputs, **kwargs)
+
+        # Run test to populate FusionCache. Deep copy inputs for this run but
+        # not the final run. When a fusion output aliases an input, it will
+        # change the input value for subsequent function calls. Therefore, only
+        # the final run should take the original tensors and potentially update
+        # their values.
+        inputs_copy = deepcopy(inputs)
+        test_fn(self, fusion_func, inputs_copy, **kwargs)
+
+        # Serialize and Deserialize FusionCache
+        basic_serde_check()
+
+        # Run test with repopulated FusionCache
+        kwargs["new_fusion_expected"] = False
+        return test_fn(self, fusion_func, inputs, **kwargs)
+
+    return inner_fn
 
 
 def setUpModule():
